requirements.txt
nicegui>=3.4.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pyyaml>=6.0.0
